refactor: renames git-primitives.js git-primitives-sync.js
